+++ src/shared/interfaces/SingInInterface.tsx
@@ -16,3 +16,17 @@ export interface SignInProps {
 		sendRecoverPasswordEmail: RecoverPasswordFunction
 	}
 }
+
+type RegisterNewUserFunction = (username: string, email: string, password: string, newsLetter: boolean) => void
+
+export interface RegisterNewUserData {
+	data: {
+		registerHeaderText: string
+		usernameText: string
+		passwordText: string
+		passwordConfirmText: string
+		emailText: string
+		passwordMismatchedText: string
+		registerNewUser: RegisterNewUserFunction
+	}
+}
