+++ src/view/signin/SignInView.tsx
@@ -1,7 +1,7 @@
 import React, { useContext, useState } from 'react'
 import { useHistory } from 'react-router'
 import styled from 'styled-components'
-import { RegisterNewUserData, RegisterUser } from './components/RegisterUser'
+import { RegisterUser } from './components/RegisterUser'
 import { SignInUser } from './components/SignInUser'
 import RoutingPath from '../../routes/RoutingPath'
 import CodicAPIService from '../../shared/api/services/CodicAPIService'
@@ -9,7 +9,7 @@ import LocalStorage from '../../shared/cache/LocalStorage'
 import { windowsMaxWidth } from '../../shared/data/WindowsSizes'
 import initialImage from '../../shared/images/teacher2.jpg'
 import { DimensionsInterface } from '../../shared/interfaces/DimensionsInterface'
-import { SignInProps } from '../../shared/interfaces/SingInInterface'
+import { RegisterNewUserData, SignInProps } from '../../shared/interfaces/SingInInterface'
 import { LoginCredentials, RegisterNewUser } from '../../shared/interfaces/UserInterface'
 import { UserContext } from '../../shared/providers/UserProvider'
 import { primaryBackgroundColor, primaryColor, secondaryColor } from '../../shared/styles/GlobalStyle'
